requirements.txt
fastapi[standard]
sqlmodel
passlib[bcrypt]
python-jose